// frontend/src/api/generated/models/settlementPreviewResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OIDE API
 * 정산방 생성/참여, 결제 등록, 정산 계산 API
 * OpenAPI spec version: v1
 */
import type { MemberResultResponse } from './memberResultResponse';
import type { RateResponse } from './rateResponse';
import type { TransferResponse } from './transferResponse';

export interface SettlementPreviewResponse {
  settlementAvailable?: boolean;
  invalidPaymentIds?: number[];
  missingCurrencies?: string[];
  rates?: RateResponse[];
  memberResults?: MemberResultResponse[];
  transfers?: TransferResponse[];
}